feat: Add security-txt service to docker-compose.yml
The code changes add a new service called "security-txt" to the docker-compose.yml file. This service is responsible for serving a security.txt file. The environment variable "SECURITY_STRING" is set to "This is a test security string." The service is labeled and configured to work with Traefik for routing and SSL/TLS termination.

Note: The suggested commit message adheres to the conventional format of "feat: <commit message>".