feat: Update security-txt service to use latest container image
The code changes update the "security-txt" service in the docker-compose.yml file to use the latest container image from the repository "git.b0t.at/b0t-at/security.txt". This ensures that the service is running with the most up-to-date version and includes any bug fixes or improvements.

Note: The suggested commit message follows the conventional format of "feat: <commit message>" and provides a clear and concise description of the code changes.